Add Windows support [requires ensure_native()]
Add Windows support.
This requires using DrawingArea.ensure_native() to ensure that the
DrawingArea has a native window assigned.  In Windows, if this is not
done, the video is written to the parent OS window (not a "window" in
the traditional sense of an app, but rather in the window manager
clipped rectangle sense).  The symptom is that the video will be drawn
over top of any widgets, etc. in the parent window.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

See link for more details:
    http://stackoverflow.com/questions/9259731/getting-gtk-gdk-window-handle-on-windows