Add WindowServiceProxy class
The WindowServiceProxy class wraps the creation of a subprocess.Popen
instance running a WindowService server.  Since WindowServiceProxy
implements __enter__/__exit__, it can be used as a context manager,
e.g.,
    with WindowServiceProxy() as w:
        do something...

Using a with statement ensures that all server processes are properly
terminated.  Otherwise, it is up to the user to either call the close()
method, or explicitly delete the WindowServiceProxy object to ensure
cleanup is done.