Add draw_queue and with_scale args to create_pi...
Add draw_queue and with_scale args to create_pipeline() service API.

draw_queue can be set to a DrawQueue instance, allowing a set of cairo
drawing commands to be replayed on top of each video frame.

with_scale can be set to True to add support for a scale() method in the
GStreamer service API.  A call to scale(width, height) will cause the
display video branch of a pipeline to be scaled accordingly.  Note that
this scaling takes place before the cairo drawing in the pipeline.  The
reason is that this allows the cairo drawing to be scaled properly,
which is especially important when scaling a video up, since the cairo
drawing would appear blocky otherwise.